+++ packages/sitecore-jss/src/editing/utils.ts
@@ -20,7 +20,10 @@ export const PAGES_EDITING_MARKER = 'jss-hrz-editing';
 /**
  * Default allowed origins for editing requests. This is used to enforce CORS, CSP headers.
  */
-export const EDITING_ALLOWED_ORIGINS = ['https://pages.sitecorecloud.io'];
+export const EDITING_ALLOWED_ORIGINS = [
+  'https://pages.sitecorecloud.io',
+  'https://app.sitecorecloud.io',
+];
 
 type ExtendedWindow = Window &
   typeof globalThis & {